+++ app/src/main/java/com/ainsigne/caa_newsapp/navigation/NavigationCollection.kt
@@ -1,10 +1,10 @@
 package com.ainsigne.caa_newsapp.navigation
 
-
 import com.ainsigne.common.Navigation
 import com.ainsigne.common.navigation.CustomNavigation
 import com.ainsigne.common.navigation.HomeNavigation
 import com.ainsigne.common.navigation.SplashNavigation
+import com.ainsigne.domain.navigation.ArticleDetails
 import timber.log.Timber
 
 /**
@@ -68,12 +68,11 @@ class NavigationCollection {
     ) {
         when (navigation) {
             Navigation.Home(HomeNavigation.HOME_TO_DETAILS) -> {
-
+                navigationViewModel.homeToDetails(data as ArticleDetails?)
             }
             else -> {
                 Timber.d(" Unhandled Navigation $data")
             }
         }
     }
-
 }