+++ features/home/src/main/java/com/ainsigne/home/di/HomeComponent.kt
@@ -1,6 +1,5 @@
 package com.ainsigne.home.di
 
-
 import com.ainsigne.common.scopes.Fragment
 import com.ainsigne.core.di.components.CoreComponent
 import com.ainsigne.home.ui.HomeFragment