+++ features/home/src/main/java/com/ainsigne/home/ui/HomeFragment.kt
@@ -1,21 +1,22 @@
 package com.ainsigne.home.ui
 
 import android.content.Context
+import com.ainsigne.common.Navigation
 import com.ainsigne.common.base.ui.BaseFragment
+import com.ainsigne.common.navigation.HomeNavigation
 import com.ainsigne.common.utils.CANADA
 import com.ainsigne.common.utils.US
-import com.ainsigne.common.utils.extension.activity
 import com.ainsigne.common.utils.extension.setOnSingleClickListener
 import com.ainsigne.common.utils.extension.showError
 import com.ainsigne.common.utils.network.NetworkStatus
 import com.ainsigne.core.di.coreComponent
+import com.ainsigne.domain.navigation.ArticleDetails
 import com.ainsigne.home.R
 import com.ainsigne.home.databinding.FragmentHomeBinding
 import com.ainsigne.home.di.DaggerHomeComponent
 import com.ainsigne.home.ui.adapter.HeadlineAdapter
 import com.ainsigne.home.viewmodel.HomeViewModel
 import javax.inject.Inject
-
 
 class HomeFragment : BaseFragment<FragmentHomeBinding>(
     FragmentHomeBinding::inflate
@@ -25,7 +26,17 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(
     lateinit var viewModel: HomeViewModel
 
     val adapter = HeadlineAdapter {
-
+        navigationCallback?.navigateWith(
+            Navigation.Home(HomeNavigation.HOME_TO_DETAILS),
+            ArticleDetails(
+                title = it.title,
+                author = it.author,
+                description = it.description,
+                content = it.content,
+                urlToImage = it.urlToImage,
+                publishedAt = it.publishedAt
+            )
+        )
     }
 
     override fun initializeUI() {
@@ -53,7 +64,6 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(
                             R.id.button_toggle_us
                         )
                     }
-
                 }
             }
         }
@@ -87,6 +97,4 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(
         super.onAttach(context)
         DaggerHomeComponent.factory().create(coreComponent()).inject(this)
     }
-
-
 }