+++ features/splash/src/main/java/com/ainsigne/splash/ui/SplashFragment.kt
@@ -4,9 +4,9 @@ import com.ainsigne.common.Navigation
 import com.ainsigne.common.base.ui.BaseFragment
 import com.ainsigne.common.navigation.SplashNavigation
 import com.ainsigne.common.utils.extension.noRefresh
+import com.ainsigne.common.utils.placeholder
 import com.ainsigne.common.utils.ui.DelayHelper
 import com.ainsigne.splash.databinding.FragmentSplashBinding
-
 
 class SplashFragment : BaseFragment<FragmentSplashBinding>(
     FragmentSplashBinding::inflate
@@ -20,13 +20,10 @@ class SplashFragment : BaseFragment<FragmentSplashBinding>(
     }
 
     override fun initializeObservers() {
-
+        placeholder()
     }
 
     override fun initializeToBeRefresh(): () -> Unit {
         return noRefresh()
     }
-
-
-
 }