+++ features/splash/src/test/java/com/ainsigne/splash/ExampleUnitTest.kt
@@ -3,7 +3,6 @@ package com.ainsigne.splash
 import org.junit.Assert.assertEquals
 import org.junit.Test
 
-
 /**
  * Example local unit test, which will execute on the development machine (host).
  *